requirements.txt
# LangChain DeepAgents Framework
deepagents>=0.1.0

# Google Gemini LLM
langchain-google-genai>=2.0.0
# OpenRouter Support
langchain-openai>=0.1.0

# LangServe for HTTP API
langserve[all]>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0

# MCP Server/Client
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0

# LangGraph for checkpointing
langgraph>=0.4.0
langgraph-checkpoint>=3.0.0

# Utilities
python-dotenv>=1.0.0

# YAML 파싱 (AGENT.md frontmatter)
pyyaml>=6.0.0
pydantic>=2.0.0
